Refactor UI and server structure
