+++ desktop/app/src/App.tsx
@@ -4,13 +4,14 @@
 // API base comes from VITE_API_BASE (defaults to http://127.0.0.1:8000)
 
 import { useEffect, useRef, useState } from "react";
-import { createPortal } from "react-dom";
-
-// ---------- Config ----------
-const DEFAULT_API_BASE = "http://127.0.0.1:8000";
-const API_BASE =
-  (import.meta as any)?.env?.VITE_API_BASE?.toString() || DEFAULT_API_BASE;
-
+import { ActiveStrategies } from "./components/ActiveStrategies";
+import { StrategyCatalog } from "./components/StrategyCatalog";
+import { NiceCombobox, NiceSelect } from "./components/NiceSelect";
+import { useLocalStorage } from "./hooks/useLocalStorage";
+import { useToast } from "./hooks/useToast";
+import { api, GET, SEND } from "./lib/api";
+import { askConfirm, brief, nowIso, sleep } from "./lib/utils";
+import type { Mode, RiskConfig } from "./types";
 // ---------- Styles ----------
 const css = `
 :root {
@@ -101,7 +102,7 @@ body{margin:0;background:var(--bg);
   background-size: 12px 12px; background-repeat: no-repeat;
 }
 .menu {
-  position: absolute; left: 0; right: 0; z-index: 50; margin-top: 6px;
+  position: absolute; left: 0; right: 0; z-index: 10000; margin-top: 6px;
   background: rgba(16,20,31,.98);
   border: 1px solid var(--border); border-radius: 12px; padding: 6px;
   box-shadow: 0 16px 40px rgba(0,0,0,.35);
@@ -179,110 +180,8 @@ type RiskConfig = {
   flatten_before_close_min?: number;
 };
 
-type StrategyKind = "ma-crossover" | "ma-grid";
 
 // ---------- Small helpers ----------
-function useLocalStorage<T>(key: string, initial: T) {
-  const [v, setV] = useState<T>(() => {
-    try {
-      const raw = localStorage.getItem(key);
-      return raw ? (JSON.parse(raw) as T) : initial;
-    } catch { return initial; }
-  });
-  useEffect(() => {
-    try { localStorage.setItem(key, JSON.stringify(v)); } catch {}
-  }, [key, v]);
-  return [v, setV] as const;
-}
-
-async function askConfirm(message: string) {
-  const g = (window as any);
-  try {
-    if (g.__TAURI__?.dialog?.confirm) {
-      return await g.__TAURI__.dialog.confirm(message, { title: "Confirm", type: "warning" });
-    }
-  } catch {}
-  return window.confirm(message);
-}
-
-async function GET<T>(path: string, params?: Record<string, any>): Promise<T> {
-  const url = new URL(path, API_BASE);
-  if (params) Object.entries(params).forEach(([k, v]) => url.searchParams.set(k, String(v)));
-  const r = await fetch(url, { credentials: "omit" });
-  if (!r.ok) throw new Error(await r.text());
-  return r.json() as Promise<T>;
-}
-async function SEND<T>(path: string, body?: any, method: "POST" | "PUT" | "PATCH" = "POST"): Promise<T> {
-  const r = await fetch(new URL(path, API_BASE), {
-    method, headers: { "Content-Type": "application/json" },
-    body: body ? JSON.stringify(body) : undefined,
-  });
-  if (!r.ok) throw new Error(await r.text());
-  return r.json() as Promise<T>;
-}
-const sleep = (ms: number) => new Promise(res => setTimeout(res, ms));
-const nowIso = () => new Date().toLocaleTimeString();
-
-// ---------- API bindings ----------
-const api = {
-  // connection
-  connect: (host: string, port: number, client_id: number) => SEND("/connect", { host, port, client_id }),
-  accountsActive: () => GET<{ account_id: string | null; trd_env: string | null }>("/accounts/active"),
-  selectAccount: (account_id: string, trd_env: "SIMULATE" | "REAL") =>
-    SEND("/accounts/select", { account_id, trd_env }),
-
-  // session helpers
-  sessionStatus: () => GET<{ saved: any; connected: boolean; active_account: any }>("/session/status"),
-  sessionSave: (host: string, port: number, account_id?: string, trd_env?: string) =>
-    SEND("/session/save", { host, port, account_id, trd_env }),
-  sessionClear: () => SEND("/session/clear", {}),
-
-  // bot core
-  getBotMode: () => GET<{ mode: Mode }>("/bot/mode"),
-  setBotMode: (mode: Mode) => SEND<{ mode: Mode }>("/bot/mode", { mode }, "PUT"),
-  getRiskConfig: () => GET<RiskConfig>("/risk/config"),
-  putRiskConfig: (cfg: RiskConfig) => SEND<RiskConfig>("/risk/config", cfg, "PUT"),
-  getRiskStatus: () => GET<{ ok: boolean; config: RiskConfig; open_positions: number | null }>("/risk/status"),
-  getPnlToday: () => GET<{ date: string; realized_pnl: number }>("/pnl/today"),
-
-  flattenAll: (symbols?: string[]) => SEND("/positions/flatten", symbols?.length ? { symbols } : {}),
-  listStrategies: () => GET<Array<{ id: number; name: string; active: boolean; symbol: string }>>("/automation/strategies"),
-  stopStrategy: (id: number) => SEND(`/automation/stop/${id}`),
-
-  getActionLogs: (q: { limit?: number; symbol?: string; since_hours?: number }) => GET<any[]>("/logs/actions", q),
-  backtestMA: (payload: any) => SEND("/backtest/ma-crossover", payload),
-
-  // strategy starters (live)
-  startMA: (payload: any) => SEND("/automation/start/ma-crossover", payload),
-  // startGrid: (payload: any) => SEND("/automation/start/ma-grid", payload),
-  // autopilot
-  autopilotStatus: () => GET("/autopilot/status"),
-  autopilotEnable: (on: boolean) => SEND("/autopilot/enable", { on }),
-  autopilotPreview: () => SEND("/autopilot/preview", {}),
-  autopilotLogs: (limit: number) => GET("/autopilot/logs", { limit }),
-  // execution (SIM)
-  listExecOrders: (q: { symbol?: string; status?: string; limit?: number }) => GET<any[]>("/exec/orders", q),
-  cancelExecOrder: (id: string) => SEND(`/exec/orders/${id}/cancel`, {}, "POST"),
-  listExecFills: (q: { symbol?: string; limit?: number }) => GET<any[]>("/exec/fills", q),
-  // execution (SIM) – positions
-  listExecPositions: (q?: { symbol?: string; limit?: number }) => GET<any[]>("/exec/positions", q || {}),
-  flattenExecPositions: (symbols?: string[]) => SEND("/exec/flatten", symbols?.length ? { symbols } : {}),
-
-
-};
-
-// ---------- Toast ----------
-function useToast() {
-  const [msg, setMsg] = useState<string | null>(null);
-  const timer = useRef<number | null>(null);
-  function show(message: string, timeout = 2800) {
-    setMsg(message);
-    window.clearTimeout(timer.current!);
-    timer.current = window.setTimeout(() => setMsg(null), timeout);
-  }
-  return { msg, show };
-}
-
 // ---------- App ----------
 enum Tab { Settings=0, Status=1, Activity=2, Backtest=3 }
 
@@ -1047,162 +946,9 @@ async function doPreview() {
 }
 
 
-  function brief(err: any) {
-    try { const j = JSON.parse(String(err?.message || err)); return j?.detail || err?.message || String(err); }
-    catch { return err?.message || String(err); }
-  }
 }
 
 // ---------- Subcomponents ----------
-
-// ---------- Fancy dropdowns (portal-based, non-clipping) ----------
-type Opt = { value: string; label: string };
-
-function useOutsideClose<T extends HTMLElement>(open: boolean, onClose: () => void) {
-  const ref = useRef<T | null>(null);
-  useEffect(() => {
-    if (!open) return;
-    const h = (e: MouseEvent) => { if (ref.current && !ref.current.contains(e.target as Node)) onClose(); };
-    const k = (e: KeyboardEvent) => { if (e.key === "Escape") onClose(); };
-    document.addEventListener("mousedown", h, true);
-    document.addEventListener("keydown", k, true);
-    return () => { document.removeEventListener("mousedown", h, true); document.removeEventListener("keydown", k, true); };
-  }, [open, onClose]);
-  return ref;
-}
-
-function useAnchorPosition(trigger: HTMLElement | null, open: boolean, menuMaxH = 260) {
-  const [pos, setPos] = useState<{ left: number; top: number; width: number; openUp: boolean }>(
-    { left: 0, top: 0, width: 0, openUp: false }
-  );
-
-  useEffect(() => {
-    if (!open || !trigger) return;
-
-    // capture a non-null handle for the closure
-    const el: HTMLElement = trigger;
-
-    function place() {
-      const r = el.getBoundingClientRect(); // <- no null warning now
-      const width = Math.max(r.width, 160);
-      let left = Math.min(Math.max(8, r.left), window.innerWidth - width - 8);
-      let top = r.bottom + 6;
-      let openUp = false;
-
-      // open upward if not enough room below
-      if (top + menuMaxH > window.innerHeight - 8) {
-        openUp = true;
-        top = Math.max(8, r.top - 6 - menuMaxH);
-      }
-      setPos({ left, top, width, openUp });
-    }
-
-    place();
-    const opts: AddEventListenerOptions = { passive: true };
-    window.addEventListener("scroll", place, true);
-    window.addEventListener("resize", place, opts);
-    return () => {
-      window.removeEventListener("scroll", place, true);
-      window.removeEventListener("resize", place);
-    };
-  }, [open, trigger, menuMaxH]);
-
-  return pos;
-}
-
-export function NiceSelect({
-  value, onChange, options, width = 160, placeholder = "Select…"
-}: { value: string; onChange: (v: string) => void; options: Opt[]; width?: number; placeholder?: string }) {
-  const [open, setOpen] = useState(false);
-  const wrapRef = useRef<HTMLDivElement | null>(null);
-  const btnRef = useRef<HTMLButtonElement | null>(null);
-  const outsideRef = useOutsideClose<HTMLDivElement>(open, () => setOpen(false));
-  const pos = useAnchorPosition(btnRef.current ?? null, open);
-  const current = options.find(o => o.value === value);
-
-  return (
-    <div className="custom-select" ref={wrapRef} style={{ width }}>
-      <button className="custom-trigger" ref={btnRef} type="button" onClick={() => setOpen(o => !o)}>
-        {current?.label ?? <span style={{ color: "var(--muted)" }}>{placeholder}</span>}
-      </button>
-
-      {open && createPortal(
-        <div
-          ref={outsideRef}
-          className="menu"
-          style={{
-            position: "fixed",
-            left: pos.left,
-            top: pos.top,
-            width: pos.width,
-            maxHeight: 260,
-            zIndex: 10000,
-          }}
-        >
-          {options.map(o => (
-            <div
-              key={o.value}
-              className={`item ${o.value === value ? "active" : ""}`}
-              onClick={() => { onChange(o.value); setOpen(false); }}
-            >
-              {o.label}
-            </div>
-          ))}
-        </div>,
-        document.body
-      )}
-    </div>
-  );
-}
-
-export function NiceCombobox({
-  value, onChange, options, width = 200, placeholder = "Search…"
-}: { value: string; onChange: (v: string) => void; options: Opt[]; width?: number; placeholder?: string }) {
-  const [open, setOpen] = useState(false);
-  const [q, setQ] = useState("");
-  const btnRef = useRef<HTMLButtonElement | null>(null);
-  const outsideRef = useOutsideClose<HTMLDivElement>(open, () => setOpen(false));
-  const pos = useAnchorPosition(btnRef.current, open);
-  const current = options.find(o => o.value === value);
-  const filtered = q
-    ? options.filter(o => o.label.toLowerCase().includes(q.toLowerCase()) || o.value.toLowerCase().includes(q.toLowerCase()))
-    : options;
-
-  return (
-    <div className="custom-select" style={{ width }}>
-      <button className="custom-trigger" ref={btnRef} type="button" onClick={() => setOpen(o => !o)}>
-        {current?.label ?? <span style={{ color: "var(--muted)" }}>Select…</span>}
-      </button>
-
-      {open && createPortal(
-        <div
-          ref={outsideRef}
-          className="menu"
-          style={{
-            position: "fixed",
-            left: pos.left,
-            top: pos.top,
-            width: pos.width,
-            maxHeight: 260,
-            zIndex: 10000,
-          }}
-        >
-          <input className="search" autoFocus value={q} onChange={e => setQ(e.target.value)} placeholder={placeholder} />
-          {filtered.length ? filtered.map(o => (
-            <div
-              key={o.value}
-              className={`item ${o.value === value ? "active" : ""}`}
-              onClick={() => { onChange(o.value); setOpen(false); setQ(""); }}
-            >
-              {o.label}
-            </div>
-          )) : <div className="item" style={{ color: "var(--muted)" }}>No matches</div>}
-        </div>,
-        document.body
-      )}
-    </div>
-  );
-}
 
 
 function ActivityLog(props: {
@@ -1279,242 +1025,7 @@ function ActivityLog(props: {
   }
 }
 
-function ActiveStrategies({ onStopped, refreshKey }: { onStopped?: () => void; refreshKey?: number }) {
-  const [items, setItems] = useState<any[] | null>(null);
-  const [loading, setLoading] = useState(false);
-
-  useEffect(()=>{ refresh(); }, []);           // initial load
-  useEffect(()=>{ refresh(); }, [refreshKey]); // reload after Kill Switch
-
-  async function refresh() {
-    try {
-      setLoading(true);
-      const ls = await GET<any[]>("/automation/strategies");
-      setItems(ls || []);
-    } finally {
-      setLoading(false);
-    }
-  }
-
-  async function stop(id: number) {
-    try {
-      await SEND(`/automation/stop/${id}`, {}, "POST");
-      onStopped?.();  // this will bump stratRefreshTick in App
-      refresh();
-    } catch {}
-  }
-
-  return (
-    <div className="stack">
-      <div className="row">
-        <button className="btn" onClick={refresh}>{loading?"Refreshing…":"Refresh"}</button>
-      </div>
-      <div className="table-wrap">
-        <table>
-          <thead><tr>{["id","name","symbol","active","actions"].map(h=><th key={h}>{h}</th>)}</tr></thead>
-          <tbody>
-            {(items?.length?items:[]).map((s:any)=>(
-              <tr key={s.id}>
-                <td>{s.id}</td><td>{s.name}</td><td>{s.symbol}</td>
-                <td>{String(s.active)}</td>
-                <td>{s.active && <button className="btn red" onClick={()=>stop(Number(s.id))}>Stop</button>}</td>
-              </tr>
-            ))}
-            {!items?.length && <tr><td>No strategies</td></tr>}
-          </tbody>
-        </table>
-      </div>
-    </div>
-  );
-}
-
 // ---- Strategies Catalog (select / configure / presets / start) ----
-function StrategyCatalog({ connected }: { connected: boolean }) {
-  const [kind, setKind] = useLocalStorage<StrategyKind>("strat.kind", "ma-crossover");
-
-  // shared
-  const [symbol, setSymbol] = useLocalStorage("strat.symbol", "US.AAPL");
-
-  // ma-crossover params
-  const [fast, setFast] = useLocalStorage("strat.ma.fast", 20);
-  const [slow, setSlow] = useLocalStorage("strat.ma.slow", 50);
-  const [ktype, setKType] = useLocalStorage("strat.ma.ktype", "K_1M");
-  const [interval, setIntervalSec] = useLocalStorage("strat.ma.interval", 15);
-  const [qty, setQty] = useLocalStorage("strat.ma.qty", 1.0);
-  const [sizeMode, setSizeMode] = useLocalStorage<"shares"|"usd">("strat.ma.sizeMode", "shares");
-  const [dollarSize, setDollarSize] = useLocalStorage("strat.ma.dollar", 0.0);
-  const [allowReal, setAllowReal] = useLocalStorage("strat.ma.allowReal", false);
-
-  // grid (placeholder fields)
-  const [gridFast, setGridFast] = useLocalStorage("strat.grid.fast", 20);
-  const [gridSlow, setGridSlow] = useLocalStorage("strat.grid.slow", 50);
-
-  // presets (localStorage-based)
-  const PRESETS_KEY = "strat.presets.v1";
-  const [presets, setPresets] = useState<Record<string, any>>(() => {
-    try { return JSON.parse(localStorage.getItem(PRESETS_KEY) || "{}"); } catch { return {}; }
-  });
-  function savePresets(next: Record<string, any>) {
-    setPresets(next);
-    try { localStorage.setItem(PRESETS_KEY, JSON.stringify(next)); } catch {}
-  }
-  function makePresetPayload() {
-    if (kind === "ma-crossover") {
-      return { kind, symbol, fast, slow, ktype, interval_sec: interval, qty, size_mode: sizeMode, dollar_size: dollarSize, allow_real: allowReal };
-    } else {
-      return { kind, symbol, fast: gridFast, slow: gridSlow };
-    }
-  }
-  function onSavePreset() {
-    const name = prompt("Preset name?");
-    if (!name) return;
-    const next = { ...presets, [name]: makePresetPayload() };
-    savePresets(next);
-  }
-  function onLoadPreset(name: string) {
-    const p = presets[name];
-    if (!p) return;
-    setKind(p.kind as StrategyKind);
-    setSymbol(p.symbol || symbol);
-    if (p.kind === "ma-crossover") {
-      setFast(p.fast ?? fast); setSlow(p.slow ?? slow); setKType(p.ktype ?? ktype);
-      setIntervalSec(p.interval_sec ?? interval); setQty(p.qty ?? qty);
-      setSizeMode((p.size_mode as any) ?? sizeMode); setDollarSize(p.dollar_size ?? dollarSize);
-      setAllowReal(!!p.allow_real);
-    } else {
-      setGridFast(p.fast ?? gridFast); setGridSlow(p.slow ?? gridSlow);
-    }
-  }
-  function onDeletePreset(name: string) {
-    const next = { ...presets }; delete next[name]; savePresets(next);
-  }
-
-  async function startStrategy() {
-    if (!connected) { alert("Not connected"); return; }
-    if (kind === "ma-crossover") {
-      const payload = {
-        symbol, fast, slow, ktype,
-        qty, size_mode: sizeMode, dollar_size: dollarSize,
-        interval_sec: interval, allow_real: allowReal,
-      };
-      try { await api.startMA(payload); alert("MA strategy started."); }
-      catch (e:any) { alert(`Start failed: ${brief(e)}`); }
-    } else {
-      alert("Grid live start requires a /automation/start/ma-grid endpoint. (Backtest supports grid.)");
-    }
-  }
-
-  function brief(err: any) {
-    try { const j = JSON.parse(String(err?.message || err)); return j?.detail || err?.message || String(err); }
-    catch { return err?.message || String(err); }
-  }
-
-  return (
-    <div className="panel">
-      <h2 style={{marginTop:0,marginBottom:10}}>Strategies</h2>
-      <div className="form-row">
-        <div>
-          <div className="label">Strategy</div>
-          <NiceSelect
-            value={kind}
-            onChange={(v)=>setKind(v as StrategyKind)}
-            options={[
-              { value: "ma-crossover", label: "MA Crossover" },
-              { value: "ma-grid", label: "MA Grid (beta)"},
-            ]}
-            width={200}
-          />
-        </div>
-        <div>
-          <div className="label">Symbol</div>
-          <input className="input" value={symbol} onChange={e=>setSymbol(e.target.value)} />
-        </div>
-        <div style={{display:"flex", alignItems:"flex-end", gap:12}}>
-          <button className="btn brand" onClick={startStrategy} disabled={!connected}>Start</button>
-          <button className="btn" onClick={onSavePreset}>Save Preset</button>
-          <PresetPicker presets={presets} onLoad={onLoadPreset} onDelete={onDeletePreset}/>
-        </div>
-      </div>
-
-      {kind === "ma-crossover" ? (
-        <>
-          <div className="form-row" style={{marginTop:12}}>
-            <div><div className="label">Fast MA</div><input className="input" type="number" value={fast} onChange={e=>setFast(parseInt(e.target.value)||1)} /></div>
-            <div><div className="label">Slow MA</div><input className="input" type="number" value={slow} onChange={e=>setSlow(parseInt(e.target.value)||2)} /></div>
-            <div><div className="label">KType</div>
-              <NiceCombobox
-                value={ktype}
-                onChange={setKType}
-                options={["K_1M","K_5M","K_15M","K_30M","K_60M","K_1D"].map(k=>({value:k,label:k}))}
-                width={180}
-                placeholder="Search ktype…"
-              />
-            </div>
-          </div>
-          <div className="form-row">
-            <div><div className="label">Interval (sec)</div><input className="input" type="number" value={interval} onChange={e=>setIntervalSec(parseInt(e.target.value)||1)} /></div>
-            <div><div className="label">Qty (shares)</div><input className="input" type="number" value={qty} onChange={e=>setQty(parseFloat(e.target.value)||0)} /></div>
-            <div>
-              <div className="label">Size Mode</div>
-              <NiceSelect
-                value={sizeMode}
-                onChange={(v)=>setSizeMode(v as any)}
-                options={[{value:"shares",label:"shares"},{value:"usd",label:"usd"}]}
-                width={140}
-              />
-            </div>
-          </div>
-          <div className="form-row">
-            <div><div className="label">Dollar Size</div><input className="input" type="number" value={dollarSize} onChange={e=>setDollarSize(parseFloat(e.target.value)||0)} /></div>
-            <div>
-              <div className="label">Allow Real Trading</div>
-              <NiceSelect
-                value={String(allowReal)}
-                onChange={(v)=>setAllowReal(v==="true")}
-                options={[{value:"false",label:"False"},{value:"true",label:"True"}]}
-                width={140}
-              />
-            </div>
-          </div>
-        </>
-      ) : (
-        <>
-          <div className="form-row" style={{marginTop:12}}>
-            <div><div className="label">Fast MA</div><input className="input" type="number" value={gridFast} onChange={e=>setGridFast(parseInt(e.target.value)||1)} /></div>
-            <div><div className="label">Slow MA</div><input className="input" type="number" value={gridSlow} onChange={e=>setGridSlow(parseInt(e.target.value)||2)} /></div>
-          </div>
-          <div className="help" style={{marginTop:4}}>
-            Live Grid start requires a server endpoint <code>/automation/start/ma-grid</code>. Backtest is available in the Backtest tab.
-          </div>
-        </>
-      )}
-    </div>
-  );
-}
-
-function PresetPicker({ presets, onLoad, onDelete }:{ presets: Record<string, any>, onLoad:(n:string)=>void, onDelete:(n:string)=>void }) {
-  const names = Object.keys(presets);
-  const [sel, setSel] = useState(names[0] || "");
-  useEffect(()=>{ if (!names.includes(sel)) setSel(names[0] || ""); }, [JSON.stringify(names)]);
-  if (!names.length) return <span className="help">No presets yet.</span>;
-  return (
-    <div className="row" style={{alignItems:"flex-end"}}>
-      <div>
-        <div className="label">Presets</div>
-        <NiceCombobox
-          value={sel}
-          onChange={setSel}
-          options={names.map(n=>({value:n,label:n}))}
-          width={180}
-          placeholder="Search presets…"
-        />
-      </div>
-      <button className="btn" onClick={()=>onLoad(sel)} disabled={!sel}>Load</button>
-      <button className="btn red" onClick={()=>onDelete(sel)} disabled={!sel}>Delete</button>
-    </div>
-  );
-}
-
 function BacktestPanel() {
   const [symbol, setSymbol] = useState("US.AAPL");
   const [fast, setFast] = useState(20);
