+++ desktop/app/src/hooks/useLocalStorage.ts
@@ -1,0 +1,19 @@
+import { useEffect, useState } from "react";
+
+/** Persist state in localStorage */
+export function useLocalStorage<T>(key: string, initial: T) {
+  const [v, setV] = useState<T>(() => {
+    try {
+      const raw = localStorage.getItem(key);
+      return raw ? (JSON.parse(raw) as T) : initial;
+    } catch {
+      return initial;
+    }
+  });
+  useEffect(() => {
+    try {
+      localStorage.setItem(key, JSON.stringify(v));
+    } catch {}
+  }, [key, v]);
+  return [v, setV] as const;
+}
