+++ requirements.txt
@@ -6,6 +6,5 @@ pydantic>=1.10.0
 python-dotenv>=1.0.0
 requests>=2.31.0
 sqlmodel>=0.0.12
-streamlit>=1.21.0
 uvicorn>=0.23.0
 yfinance>=0.2.40
